requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
python-dotenv==0.19.0
openai==1.3.7
psycopg2-binary==2.9.9
SQLAlchemy==1.4.23
gunicorn==20.1.0
